# public/api.yaml
openapi: 3.0.0
info:
  title: Frutable API
  description: Frutable project
  version: 0.1.9
servers:
  - url: http://localhost:5000/api/v1
    description: Local server
  - url: http://staging-api.example.com
    description: Live Server
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []  
paths:
  /users/useradd:
    post:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                avtar:
                  type: string
                  format: binary
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /users/login:
    post:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /users/logout:
    post:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                _id:
                  type: string
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /users/get-newtoken:
    post:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /users/googlelogin:
    get:
      summary: google login
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /users/facebooklogin:
    get:
      summary: facebooklogin login
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /categories/list-categories:
    get:
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
      summary: list-categories login
      description: Optional extended description in CommonMark or HTML.
      tags:
        - Users
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
  /categories/updateCategories/{category_id}:
    put:
      summary: Categories.
      description: Optional extended description in CommonMark or HTML.
      tags:
        - categories  
      parameters:
        - in: path
          name: category_id
          schema:
            type: string
          required: true
          description: Update category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string     
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string  